+++ src/app/app.routes.ts
@@ -40,6 +40,7 @@ import { CardsDemoComponent } from './cards/cards.demo';
 import { CardsExpandableDemoComponent } from './cards/cards-expandable.demo';
 import { CardsMultiSelectDemoComponent } from './cards/cards-multi-select.demo';
 import { CardsSingleSelectDemoComponent } from './cards/cards-single-select.demo';
+import { CardsWorkspaceWidgetsComponent } from './cards/cards-workspace-widgets.demo';
 import { ChartDemoComponent } from './chart/chart.demo';
 import { CheckBoxDemoComponent } from './checkbox/checkbox.demo';
 import { CirclepagerDemoComponent } from './circlepager/circlepager.demo';
@@ -143,6 +144,7 @@ import { MaskDemoComponent } from './mask/mask.demo';
 import { MenuButtonDemoComponent } from './menu-button/menu-button.demo';
 import { MessageDemoComponent } from './message/message.demo';
 import { ModalDialogDemoComponent } from './modal-dialog/modal-dialog.demo';
+import { ModuleNavDemoComponent } from './module-nav/module-nav.demo';
 import { MonthViewDemoComponent } from './monthview/monthview.demo';
 import { MonthViewDemoInPageComponent } from './monthview/monthview-inpage.demo';
 import { NotificationDemoComponent } from './notification/notification.demo';
@@ -228,7 +230,9 @@ import { DataGridSummaryRowDemoComponent } from './datagrid/datagrid-summary-row
 import { DropdownMultiselectAttributesDemoComponent } from './dropdown/dropdown-multiselect-attributes-demo';
 import { DataGridVerticalScrollDemoComponent } from './datagrid/datagrid-vertical-scroll-to-end.demo';
 import { LookupSourceDemoComponent } from './lookup/lookup-source.demo';
+import { ImagesExampleDemoComponent } from './images/images-example.demo';
 import { InputClearableDemoComponent } from './input/input-clearable.demo';
+import { ProcessIndicatorDemoComponent } from './processindicator/process-indicator.demo';
 import { ListViewContextDemoComponent } from './listview/listview-context.demo';
 import { HeaderDefaultContentDemoComponent } from './header/header-default-content.demo';
 import { HomePageEditableFilledComponent } from './homepage/homepage-editable-filled.demo';
@@ -284,6 +288,7 @@ export const routes: Routes = [
   { path: 'cards-expandable', component: CardsExpandableDemoComponent },
   { path: 'cards-multi-select', component: CardsMultiSelectDemoComponent },
   { path: 'cards-single-select', component: CardsSingleSelectDemoComponent },
+  { path: 'cards-workspace-widgets', component: CardsWorkspaceWidgetsComponent },
   { path: 'chart', component: ChartDemoComponent },
   { path: 'checkbox', component: CheckBoxDemoComponent },
   { path: 'circlepager', component: CirclepagerDemoComponent },
@@ -388,6 +393,7 @@ export const routes: Routes = [
   { path: 'homepage-scenario-c', component: HomePageScenarioCDemoComponent },
   { path: 'hyperlink', component: HyperlinkDemoComponent },
   { path: 'icon', component: IconDemoComponent },
+  { path: 'images-example', component: ImagesExampleDemoComponent },
   { path: 'input-clearable', component: InputClearableDemoComponent },
   { path: 'label', component: LabelDemoComponent },
   { path: 'line', component: LineDemoComponent },
@@ -405,6 +411,7 @@ export const routes: Routes = [
   { path: 'menu-button', component: MenuButtonDemoComponent },
   { path: 'message', component: MessageDemoComponent },
   { path: 'modal-dialog', component: ModalDialogDemoComponent },
+  { path: 'module-nav', component: ModuleNavDemoComponent },
   { path: 'monthview', component: MonthViewDemoComponent },
   { path: 'monthview-inpage', component: MonthViewDemoInPageComponent },
   { path: 'monthview-legend', component: MonthViewLegendDemoComponent },
@@ -417,6 +424,7 @@ export const routes: Routes = [
   { path: 'popdown', component: PopDownDemoComponent },
   { path: 'popupmenu', component: PopupMenuDemoComponent },
   { path: 'progress', component: ProgressDemoComponent },
+  { path: 'process-indicator', component: ProcessIndicatorDemoComponent },
   { path: 'radar', component: RadarDemoComponent },
   { path: 'radiobutton', component: RadioButtonDemoComponent },
   { path: 'rating', component: RatingDemoComponent },